isep_courses_adapt/migration_scripts/requirements.txt
SQLAlchemy==1.3.22
pyodbc==4.0.30
pandas==1.2.0
python-dotenv==0.15.0
mysql-connector-python==8.0.22